doc: fix typo and rst style fixes
As per the official docs¹, we should use ``code`` instead of :samp:`code`,
since we are not enhancing variables inside.

:samp:
A piece of literal text, such as code. Within the contents, you can use curly
braces to indicate a “variable” part, as in file. For example, in :samp:`print
1+{variable}`, the part variable would be emphasized.

If you don’t need the “variable part” indication, use the standard ``code``
instead.

¹: https://www.sphinx-doc.org/en/1.5/markup/inline.html#role-samp